feat: implement HistoryCards component
